move source search out of app into a separate component
